Moved clutter to extension class for servicecollection
